requirements.txt
beautifulsoup4
pillow
setuptools